Move the generate outline shader method into outline shader text. Because this shader has non-visual test.
